% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_dist.R
\name{arch_dist}
\alias{arch_dist}
\title{Compute distance to archetypes}
\usage{
arch_dist(data, archetypes, dist_metric = "euclidean")
}
\arguments{
\item{data}{matrix of dim(variables/dimentions, examples)}

\item{archetypes}{matrix of dim(variables/dimentions, archetypes)}

\item{dist_metric}{distance metric, currently only "euclidean" implemented}
}
\value{
matrix of distances to archetypes of dim(examples, archetypes)
}
\description{
\code{arch_dist()} calculates Euclidean distance from every point to every archetype give matrices that contain this data.
}
\examples{
# Triangle with sides of 3,4,5 units length
arch_dist(matrix(c(0,0),1,2), matrix(c(3,4),1,2))
# Random data that fits into the triangle
set.seed(4355)
archetypes = generate_arc(arc_coord = list(c(5, 0), c(-10, 15), c(-30, -20)),
                          mean = 0, sd = 1, N_dim = 2)
data = generate_data(archetypes, N_examples = 1e4, jiiter = 0.04, size = 0.9)
# Find Euclidean distance between data points and archetypes
distance = arch_dist(data, archetypes)
# Find Euclidean distance between archetypes
arc_distance = arch_dist(archetypes, archetypes)
}
\seealso{
\code{\link[ParetoTI]{fit_pch}}, \code{\link[ParetoTI]{generate_arc}}, \code{\link[ParetoTI]{generate_data}}
}
\author{
Vitalii Kleshchevnikov
}
