% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_data_table.R
\name{.archLines}
\alias{.archLines}
\title{add a line between all archetypes}
\usage{
.archLines(arc_data, arc_lab = "archetypes", pch_fit = TRUE,
  type = c("average", "all")[1], average_func = mean)
}
\arguments{
\item{arc_data}{data.table that contains positions of archetypes}

\item{arc_lab}{labels for archetypes}

\item{pch_fit}{logical, is this a single archetype fit (TRUE) or multiple ("b_pch_fit", "k_pch_fit")?}

\item{type}{average across arc_lab (average)?}
}
\description{
add a line between all archetypes
}
