% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_deriv.R
\name{find_gam_deriv}
\alias{find_gam_deriv}
\alias{summary.gam_deriv}
\alias{plot.gam_deriv}
\title{Find derivative of smooth curves from gam model fit (Generalised Additive Model, mgcv)}
\usage{
find_gam_deriv(gam_fit, n_points = 200, x = NULL, N_smooths = 9,
  weights = list(rep(1, 200), sqrt(seq(1, 0, length.out = 200)), 1/(1 +
  seq(1, 0, length.out = 200)^3))[[1]], d = 1/n_points,
  return_gam = FALSE)

\method{summary}{gam_deriv}(derivs)

\method{plot}{gam_deriv}(derivs, features = derivs$derivs$y_name,
  title = "First derivative of GAM model: gene expression = function(distance from archetype)")
}
\arguments{
\item{gam_fit}{a fitted gam object as produced by gam().}

\item{n_points}{number of points at which to evaluate derivative}

\item{x}{numeric vector where to evaluate derivatives}

\item{N_smooths}{number of bases used to represent the smooth term (\link[mgcv]{s}), 4 for cubic splines}

\item{weights}{how to weight points along x axis when calculating mean (integral) probability}

\item{d}{numeric vector (1L), finite difference interval}

\item{return_gam}{return gam model as well? By default only only the first derivative at n_points}
}
\value{
\code{find_gam_deriv()} list (S3 object, find_gam_deriv) containing function call, a data.table with values of the first derivative, GAM model fit summary (gam_sm) and (optionally) gam model fit; summary entry is NA.
}
\description{
\code{find_gam_deriv()} Finds a derivative with standard errors of smooth curves from gam model fit by finite differencing. Crucially it computes the probability that derivative is negative at each point.
}
\details{
Plot method uses ggplot() to show mean derivative and confidence intervals (2 SD from the mean). This code is a modification of last example in \link[mgcv]{predict.gam}.
}
