% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_set_activity.R
\name{find_set_activity_AUCell}
\alias{find_set_activity_AUCell}
\alias{find_set_activity_pseudoinv}
\title{Find activity of a gene set using AUCell model (continuos and binary)}
\usage{
find_set_activity_AUCell(expr_mat, assay_name = "logcounts",
  aucMaxRank = nrow(expr_mat) * 0.05, gene_sets, gene_col = "ALIAS",
  set_id_col = "GOALL", set_name_col = "TERM", binary = FALSE,
  nCores = 1, plotHist = FALSE, ...)

find_set_activity_pseudoinv(expr_mat, assay_name = "logcounts",
  gene_sets, gene_col = "ALIAS", set_id_col = "GOALL",
  set_name_col = "TERM", noself = FALSE)
}
\arguments{
\item{expr_mat}{expression matrix (genes in rows, cells in columns) or one of: dgCMatrix, ExpressionSet, and SummarizedExperiment or SingleCellExperiment both of which require assay_name.}

\item{assay_name}{name of assay in SummarizedExperiment or SingleCellExperiment, normally counts or logcounts}

\item{aucMaxRank}{argument for \link[AUCell]{AUCell_calcAUC}. Threshold to calculate the AUC.In a simplified way, the AUC value represents the fraction of genes, within the top X genes in the ranking, that are included in the signature. The parameter 'aucMaxRank' allows to modify the number of genes (maximum ranking) that is used to perform this computation. By default, it is set to 5% of the total number of genes in the rankings. Common values may range from 1 to 20%.}

\item{gene_sets}{data.table or coercible to data.table that contains gene set annotations.}

\item{gene_col}{column in gene_sets storing gene identifiers.}

\item{set_id_col}{column in gene_sets storing set identifiers.}

\item{set_name_col}{column in gene_sets storing readable set names.}

\item{binary}{binarise gene set activities using \link[AUCell]{AUCell_exploreThresholds}?}

\item{nCores}{number of cores for parallel processing. See AUCell docs for details.}

\item{plotHist}{plot the AUC histograms? \link[AUCell]{AUCell_exploreThresholds}.}

\item{...}{other arguments passed to \link[AUCell]{AUCell_exploreThresholds}.}

\item{noself}{Remove self-interactions from set annotations (when sets are TF targets)}
}
\value{
\code{find_set_activity_AUCell()} data.table of gene set activities with cell in rows and gene sets in columns. Column titled "cells" contains cell ids (column names of expr_mat).
}
\description{
\code{find_set_activity_AUCell()} finds activity of each gene set in each cell by combining AUCell functions into a pipeline. Aerts lab that developed AUCell recommends adjusting the threshold when binarising activities (Details: https://www.bioconductor.org/packages/devel/bioc/vignettes/AUCell/inst/doc/AUCell.html#determine-the-cells-with-the-given-gene-signatures-or-active-gene-sets). See documentations for details: \link[AUCell]{AUCell_buildRankings}, \link[AUCell]{AUCell_calcAUC}, \link[AUCell]{AUCell_exploreThresholds}.

\code{find_set_activity_pseudoinv()} finds activity of each gene set in each cell by solving this matrix equation: expression = activities %*% gene_assignment_to_sets => activities = pseudoinverse(gene_assignment_to_sets) %*% expression. Based on code from Inferelator package by Richard Bonneau lab.
}
