% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_deriv.R
\name{gam_deriv}
\alias{gam_deriv}
\alias{plot.gam_deriv}
\title{Find derivative of smooth curves from gam model fit (Generalised Additive Model, mgcv)}
\usage{
gam_deriv(gam_fit, n_points = 200, x = NULL, N_smooths = 9,
  weights = c(1, sqrt(seq(1, 0, length.out = 200)), 1/(1 + seq(1, 0,
  length.out = 200)^3))[1], d = 1/n_points)

\method{plot}{gam_deriv}(derivs)
}
\arguments{
\item{gam_fit}{a fitted gam object as produced by gam().}

\item{n_points}{number of points at which to evaluate derivative}

\item{x}{numeric vector where to evaluate derivatives}

\item{N_smooths}{number of curves for each predictor in the model}

\item{weights}{how to weight points along x axis when calculating mean (integral) probability}

\item{d}{numeric vector (1L), finite difference interval}
}
\value{
\code{gam_deriv()} list (S3 object, gam_deriv) containing data.table-s for each predictor in gam model
}
\description{
\code{gam_deriv()} Finds a derivative with standard errors of smooth curves from gam model fit by finite differencing. Crucially it computes the probability that derivative is negative at each point.
}
\details{
Plot method uses ggplot() to show mean derivative and confidence intervals (2 SD from the mean). This code is a modification of last example in \link[mgcv]{predict.gam}.
}
