% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_arc.R
\name{generate_arc}
\alias{generate_arc}
\alias{generate_data}
\title{Generate random data within specified archetypes}
\usage{
generate_arc(arc_coord = list(c(5, 0), c(-10, 15), c(-30, -20)),
  mean = 0, sd = 1, N_dim = 2)

generate_data(archetypes, N_examples = 10000, jiiter = 0.1, size = 1)
}
\arguments{
\item{arc_coord}{list of archetype coordinates, one numeric vector (length of N dimensions) per each archetype.}

\item{mean}{mean of random distribution added to arc_coord}

\item{sd}{standard deviationn of random distribution added to arc_coord}

\item{N_dim}{number of dimentions. Can be set to large number of dimentions, in that case arc_coord will be recycled keeping true dimensionality of the data as specified in arc_coord.}

\item{archetypes}{matrix of archetypes of dim(dimensions, archetypes)}

\item{N_examples}{number of examples to be generated}

\item{jiiter}{add noise to weigth so that data is not a perfect polytope (e.g. triangle, see examples)}

\item{size}{scale the data within a polytope}
}
\value{
\code{generate_arc()} object of class "random_arc" (similar to "pch_fit"), element XC is a matrix of archetypes of dim(dimensions, archetypes)

\code{generate_data()} matrix of archetypes of dim(dimensions, examples)
}
\description{
\code{generate_arc()} generates the matrix of archetypes using specified coordinates (some noise added).

\code{generate_data()} produces matrix of random data that fits a polytope defined by archetypes by multiplying position of archetypes by random weigths (that sum to 1)
}
\examples{
# Random data that fits into the triangle
set.seed(4355)
archetypes = generate_arc(arc_coord = list(c(5, 0), c(-10, 15), c(-30, -20)),
                          mean = 0, sd = 1, N_dim = 2)
data = generate_data(archetypes$XC, N_examples = 1e4, jiiter = 0.04, size = 0.9)
# Find Euclidian distance between data points and archetypes
distance = sqrt(arch_dist(data, archetypes))
# Find Euclidian distance between archetypes
arc_distance = sqrt(arch_dist(archetypes, archetypes))
}
\seealso{
\code{\link[ParetoTI]{fit_pch}}, \code{\link[ParetoTI]{arch_dist}}
}
\author{
Vitalii Kleshchevnikov
}
