% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_sketch.R
\name{geo_sketch}
\alias{geo_sketch}
\title{Create downsampled representation of the data with Geometric Sketch}
\usage{
geo_sketch(data, N, assay_slot = "logcounts", use_PCs = TRUE,
  PCs = 100, k = "auto", seed = 5232, replace = FALSE,
  alpha = 0.1, max_iter = 200, verbose = 0)
}
\arguments{
\item{data}{matrix, sparse matrix or SingleCellExperiment (SummarizedExperiment): dim(dimensions * examples)}

\item{N}{number of cells to sample}

\item{assay_slot}{slot in data (SingleCellExperiment, SummarizedExperiment) containing matrix to use for PCA and geometric sampling}

\item{use_PCs}{logical, use PCs (TRUE) or data directly (FALSE)?}

\item{PCs}{number of PCs to use. Identified using Facebook implementation of PCA (fbpca).}

\item{k}{Number of covering boxes. When `'auto'` and replace is `True`, draws sqrt(X.shape[0]) covering boxes. When `'auto'` and replace is `False`, draws N covering boxes.}

\item{seed}{Random number generation seed passed to numpy}

\item{alpha}{Binary search halts when it obtains between `k * (1 - alpha)` and `k * (1 + alpha)` covering boxes.}

\item{max_iter}{Maximum iterations at which to terminate binary seach in rare case of non-monotonicity of covering boxes with box side length.}

\item{verbose}{report progress}
}
\value{
\code{geo_sketch()} integer vector of cell indices
}
\description{
\code{geo_sketch()} Create a downsampled representation of the data while preserving the structure and rare populations with Geometric Sketch method (https://github.com/brianhie/geosketch, https://www.biorxiv.org/content/10.1101/536730v1). This method splits gene expression space into hypercubes (covering boxes) of equal volume and samples uniformly from those cubes. This reduces the density of the data and cell numbers while preserving global structure. Useful for fitting polytopes to large datasets. First step - find PCs using Facebook PCA method, second step - define hypercubes and sample cells.
}
