% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_annot.R
\name{map_gene_annot}
\alias{map_gene_annot}
\title{Convert gene identifiers with AnnotationHub}
\usage{
map_gene_annot(taxonomy_id = 9606, keys = "TP53",
  columns = c("ENSEMBL"), keytype = "SYMBOL", pick_one = TRUE,
  localHub = FALSE, return_org_db = FALSE,
  ann_hub_cache = AnnotationHub::getAnnotationHubOption("CACHE"))
}
\arguments{
\item{taxonomy_id}{Taxonomy id of your species. 9606 is human, 10090 is mouse. Find your species id using taxonomy search by UniProt: https://www.uniprot.org/taxonomy/.}

\item{keys}{gene or protein identifiers of \code{keytype}}

\item{columns}{columns to retrieve. Details: \link[AnnotationDbi]{colsAndKeytypes}.}

\item{keytype}{type of \code{keys} identifiers. Use \code{keytypes(map_gene_annot(taxonomy_id = 9606, return_org_db = T))} to find which keytypes are available. Most frequent options: "ENSEMBL", "SYMBOL", "ALIAS".}

\item{pick_one}{pick one mapped annotation (from columns) for each key. Works only with one column.}

\item{localHub}{set localHub = FALSE for working offline. Details \link[AnnotationHub]{AnnotationHub}.}

\item{return_org_db}{return OrgDb database link instead of gene annotations.}

\item{ann_hub_cache}{If needed, set location of local AnnotationHub cache to specific directory by providing it here. To view where cache is currently located use getAnnotationHubOption("CACHE")}
}
\value{
data.frame containing mapped identifiers
}
\description{
map_gene_annot() Maps gene annotations using OrgDb databases from AnnotationHub. Most useful for mapping gene identifiers. This function can pick one mapped annotation for each key.
}
