% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arc_var.R
\name{plot_arc_var}
\alias{plot_arc_var}
\title{Plot SSE and variance explained by polytope fit}
\usage{
plot_arc_var(arc_data, type = c("varexpl", "SSE", "res_varexpl",
  "total_var")[1], point_size = 2, line_size = 1.5)
}
\arguments{
\item{arc_data}{object of class "k_pch_fit", storing the position of vertices of best fit polytopes with different k, and other data from \code{\link[ParetoTI]{fit_pch}}() run. arc_data$XC is a list where each element is XC matrix of dim(dimensions, archetypes) storing positions of vertices with different k.}

\item{type}{which measure to plot as a function of k, one of "varexpl", "SSE", "res_varexpl"}

\item{line_size}{width of lines connecting archetypes}

\item{arch_size}{size of archetype point}
}
\value{
\code{plot_arc_var()} ggplot2 (2D) plot
}
\description{
\code{plot_arc_var()} shows SSE and variance explained by polytope models with different number of vertices (k)
}
\examples{
library(ParetoTI)
library(ggplot2)
# Random data that fits into the triangle (2D)
set.seed(4355)
archetypes = generate_arc(arc_coord = list(c(5, 0), c(-10, 15), c(-30, -20)),
                          mean = 0, sd = 1, N_dim = 2)
data = generate_data(archetypes$XC, N_examples = 1e4, jiiter = 0.04, size = 0.9)
arc_data = k_fit_pch(data, ks = 1:4, check_installed = T, delta=0.1, order_by_side = F)
# Show polytopes and the data
plot_arc(arch_data = arc_data, data = data,
         which_dimensions = 1:2, type = "all", arch_size = 2,
         colors = c("#D62728", "#1F77B4", "#2CA02C", "#17BED0", "grey")) +
  theme_bw()
# Show variance explained by a polytope with each k
plot_arc_var(arc_data, type = c("varexpl", "SSE", "res_varexpl")[1],
             point_size = 2, line_size = 1.5) + theme_bw()
}
\seealso{
\code{\link[ParetoTI]{fit_pch}}, \code{\link[ParetoTI]{plot_arc}}
}
\author{
Vitalii Kleshchevnikov
}
