% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arch_struct.R
\name{plot_arch_struct}
\alias{plot_arch_struct}
\title{Find and plot similarity structure between archetypes}
\usage{
plot_arch_struct(arc, type = c("cells", "space", "marker_genes",
  "genesXarch")[1], dist_fun = function(x) dist(x, method = "euclidean"),
  marker_genes = NULL, marker_genes_mean = F)
}
\arguments{
\item{arc}{object containing archetypal analysis results (class "pch_fit")}

\item{type}{one of c("cells", "space", "marker_genes", "genesXarch"). "cells": distances are computed over archetype weights of cells (S matrix). "space": measures distances between archetypes in space, PCs or gene expression, depending on which space was used to find archetypes (XC matrix). "marker_genes": distance is computed base on lists marker genes. "genesXarch": show both archetypes and dimensions, both clustered with hierarchical clustering. When \code{marker_genes} is provided "genesXarch" shows marker gene membership across archetypes rather than expression values of those genes (PCs) at archetypes.}

\item{dist_fun}{function use to compute distance, should take one argument and compute distances between rows.}

\item{marker_genes}{filtered data.table containing markers for each archetype, normally the output of \code{\link[ParetoTI]{get_top_decreasing}} stored in $enriched_genes. When \code{type = "genesXarch"} dimensions are filtered using gene names supplied in \code{marker_genes}.}

\item{marker_genes_mean}{show the strength of marker gene association with archetypes? (mean_diff in the output of \code{\link[ParetoTI]{get_top_decreasing}} stored in $enriched_genes). By default is FALSE - show only marker gene membership (0/1).}
}
\value{
Matrix of the same dimention as the original matrix but with values in each column permuted.
}
\description{
Find similarity structure between archetypes by comparing archetype weights for cells, measuring similarity of archetypes in expression space and in marker gene. genesXarch option shows
}
\author{
Vitalii Kleshchevnikov
}
