% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_vert.R
\name{build_logit_vert}
\alias{build_logit_vert}
\alias{make_features}
\alias{predict_vert}
\alias{split_train_val}
\alias{plot_confusion_vert}
\alias{get_feature_weights}
\title{Logistic regression model to classify vertices}
\usage{
build_logit_vert(features_list, penalty = 0.7,
  regularizer = keras::regularizer_l2, initializer = "random_uniform",
  activation = "softmax", loss = "categorical_crossentropy",
  metrics = list("accuracy"))

make_features(data_attr, features_list, n_bins = 10, n_samples = 10,
  select_bins = seq_len(n_bins), combine_bins = seq_len(n_bins))

predict_vert(vert_model, data_attr, features_data)

split_train_val(features_data, val_prop = 0.3, per_class = TRUE)

plot_confusion_vert(predicted)

get_feature_weights(vert_model, features_data)
}
\arguments{
\item{val_prop}{proportion of samples used for validation.}

\item{per_class}{Use proportion of samples within each class for validation? By default the function takes class into account when splitting data (TRUE)}
}
\description{
build_logit_vert() defines keras model

make_features() generate features for predicting vertex by expression profile of several genes with distance from that vertex

predict_vert() predict vertices using a model and attribute data (distance to vertex + feature of cells)

split_train_val() split data into training and validation. Shuffle examples to ensure that model sees all vertices in training and validation

plot_confusion_vert() plot vertex probabilities as confusion matrix

get_feature_weights() get gene weights and ranked features from keras model
}
