% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_data_table.R
\name{project}
\alias{project}
\title{Project in low dimentions (PCA) and process data and polytopes fits for plotting}
\usage{
project(arc_data, data)
}
\arguments{
\item{arc_data}{list of matrices storing the position of archetypes, class "pch_fit", "r_pch_fit". Each element of a list represents an independent run of the polytope fitting algorithm}

\item{data}{matrix of data in which archetypes/polytope were found}
}
\description{
Project in low dimentions (PCA) and process data and polytopes fits for plotting
}
