% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_to_pcs}
\alias{project_to_pcs}
\title{Project archetype and data in low dimentions (PCA)}
\usage{
project_to_pcs(arc_data = NULL, data, n_dim = 3, s = NULL)
}
\arguments{
\item{arc_data}{objects of class "pch_fit", "b_pch_fit", "k_pch_fit" storing the position of archetypes and other data produced by \code{\link[ParetoTI]{fit_pch}}(). arc_data$XC is matrix of dim(dimensions, archetypes) or list where each element is XC matrix from an independent run of the polytope fitting algorithm.}

\item{data}{matrix of data in which archetypes/polytope were found, dim(variables/dimentions, examples)}

\item{n_dim}{number of principal component dimensions}

\item{s}{list 's' containing SVD decomposition results (U, d, Vt), standard deviation and mean of genes used for decomposition (sds, means)}
}
\description{
project_to_pcs() projects archetypes (\code{arc_data}) and data points (\code{arc_data}) to PC space. Data is standardised prior to PCA by default.

project_from_pc() projects archetypes and data points to original space provided SVD decomposition results
}
\examples{
# Random data that fits into the triangle
set.seed(4355)
arc_data = generate_arc(arc_coord = list(c(5, 0), c(-10, 15), c(-30, -20)),
                          mean = 0, sd = 1)
data = generate_data(arc_data$XC, N_examples = 1e4, jiiter = 0.04, size = 0.9)
}
