% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_var.R
\name{rand_var}
\alias{rand_var}
\title{Permute each column of a matrix}
\usage{
rand_var(X)
}
\arguments{
\item{X}{matrix to be permuted}
}
\value{
matrix of the same dimention as the original matrix but with values in each column permuted
}
\description{
This function uses permutations to disrupts relationships between variables but keep the distribution of each variable constant. Variables should be in columns.
}
\examples{
set.seed(4354)
N = 1000
X = matrix(rnorm(N * 10 * N), N * 10, N)
}
\seealso{
\code{\link{}}, \code{\link{}}
}
\author{
Vitalii Kleshchevnikov
}
