% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_data_table.R
\name{.archLines}
\alias{.archLines}
\title{add a line between all archetypes}
\usage{
.archLines(arc_data, arc_lab = "archetypes", type = c("average",
  "all")[1], average_func = mean)
}
\arguments{
\item{arc_data}{data.table that contains positions of data points and archetypes}

\item{label}{character specifying which data point to connect (when multiple replicates of fitted polytopes)}
}
\description{
add a line between all archetypes
}
