% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{install_py_pcha}
\alias{install_py_pcha}
\title{Install python module py_pcha and numpy}
\usage{
install_py_pcha(method = "auto", conda = "auto",
  python_version = "python 2.7.10", envname = "reticulate_PCHA",
  overwrite_env = F)
}
\arguments{
\item{method}{paratemer for \code{\link[reticulate]{py_install}}. Default option should work in most cases. Use virtualenv if you don't want to install anaconda but virtualenv doesn't work on Windows.}

\item{conda}{paratemer for \code{\link[reticulate]{py_install}}. Default option should work in most cases.}

\item{python_version}{version to be installed into environment that is compatible with py_pcha module.}

\item{envname}{name of the conda enviroment where PCHA should be installed. If that enviroment doesn't exist it will be created. If it contains incorrect python_version the function will give an error.}

\item{overwrite_env}{It TRUE overwrites conda environment.}
}
\value{
path to python enviroment with py_pcha module installed
}
\description{
Install python module py_pcha and numpy
}
\details{
This is a helper function that install py_pcha and numpy python modules into conda environment. Unless you have a strong reason please use suggested defaults.

If installation fails with an error "Cannot fetch index base URL http://pypi.python.org/simple/" try this solution: "Older versions of pip and distribute default to http://pypi.python.org/simple, which no longer works. A solution is to install an up-to-date pip and distribute using pip install -i https://pypi.python.org/simple -U pip distribute into the virtual environment before running the rest of the build process."
}
\seealso{
\code{\link{}}, \code{\link{}}
}
\author{
Vitalii Kleshchevnikov
}
