% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam.R
\name{plot_gam}
\alias{plot_gam}
\title{Plot gam model fit for multiple genes and archetypes (Generalised Additive Model, mgcv)}
\usage{
plot_gam(gam_fit, data, y = "y", smooth.cov = "x", groupCovs = NULL,
  color = NULL, color_val = c("#868686FF", "#0073C2FF"))
}
\arguments{
\item{gam_fit}{a fitted gam object as produced by gam().}

\item{x}{numeric vector where to evaluate derivatives}

\item{N_smooths}{number of curves for each predictor in the model}

\item{d}{numeric vector (1L), finite difference interval}
}
\value{
\code{plot_gam()} list (S3 object, plot_gam) containing data.table-s for each predictor in gam model
}
\description{
\code{plot_gam()} Finds a derivative with standard errors of smooth curves from gam model fit by finite differencing. Crucially it computes the probability that derivative is negative at each point.
}
\details{
Plotting of multiple gam fits is inspired by https://stackoverflow.com/questions/49471300/gam-plots-with-ggplot
}
