% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project polytope fit and data in low dimentions (PCA)}
\usage{
project(arc_data, data, n_dim = 3, proj_matrix = NULL)
}
\arguments{
\item{arc_data}{objects of class "pch_fit", "b_pch_fit", "k_pch_fit" storing the position of archetypes and other data produced by \code{\link[ParetoTI]{fit_pch}}(). arc_data$XC is matrix of dim(dimensions, archetypes) or list where each element is XC matrix from an independent run of the polytope fitting algorithm.}

\item{data}{matrix of data in which archetypes/polytope were found, dim(variables/dimentions, examples)}

\item{n_dim}{number of principal component dimensions}

\item{proj_matrix}{crossprod(proj_matrix, data) projects data/archetypes into principal component space}
}
\description{
Project polytope fit and data in low dimentions (PCA)
}
\examples{
# Random data that fits into the triangle
set.seed(4355)
arc_data = generate_arc(arc_coord = list(c(5, 0), c(-10, 15), c(-30, -20)),
                          mean = 0, sd = 1)
data = generate_data(arc_data$XC, N_examples = 1e4, jiiter = 0.04, size = 0.9)
}
